Move browser history to separate file to fix HMR
